fix(moments): use fractional centroid in 2D central moments (drop (int) truncation)
BasicGeomoms2D::centralMom() cast the centroid to int
(`- (int)originOfX`), computing central moments about a centroid rounded
DOWN to the nearest pixel — a fixed up-to-0.5px offset. Consequences:
first central moments mu10/mu01 were nonzero (should be 0 by definition),
and every odd-order central/normalized/Hu moment (shape AND intensity-
weighted) was wrong; even-order ones were off by ~0.2%.

Root-caused via the oracle-vetting harness: for the 48x40 geomoment
fixture, mu10 = 960 = 0.5*m00 exactly, and mu20 = 368960 = 368480 +
0.25*m00 — i.e. moments about x=23 instead of 23.5. Confirmed against two
independent references (scikit-image regionprops and the analytic closed
form), which both give mu10=mu01=0.

Fix: subtract the true double-valued centroid. Affected shape + intensity
central-derived goldens in test_moments_{regression,skimage}.h regenerated
(raw moments unchanged); they now match skimage exactly.

VERIFIED: full gtest suite 697/697.

+++ src/nyx/features/2d_geomoments_basic.cpp
@@ -185,7 +185,7 @@ double BasicGeomoms2D::centralMom(const pixcloud& cloud, int p, int q)
     for (auto& pxl : cloud)
     {
         double I = INTEN(double(pxl.inten));
-        sum += I * int_pow(double(pxl.x - baseX) - (int)originOfX, p) * int_pow(double(pxl.y - baseY) - (int)originOfY, q);
+        sum += I * int_pow(double(pxl.x - baseX) - originOfX, p) * int_pow(double(pxl.y - baseY) - originOfY, q);
     }
     return sum;
 }
@@ -198,7 +198,7 @@ double BasicGeomoms2D::centralMom(const pixcloud& cloud, const reintenvec& reali
     for (size_t i = 0; i < n; i++)
     {
         auto& pxl = cloud[i];
-        sum += realintens[i] * int_pow(double(pxl.x - baseX) - (int)originOfX, p) * int_pow(double(pxl.y - baseY) - (int)originOfY, q);
+        sum += realintens[i] * int_pow(double(pxl.x - baseX) - originOfX, p) * int_pow(double(pxl.y - baseY) - originOfY, q);
     }
     return sum;
 }
